Tighten Open CryptoPay UI integration
